Update to main website

+++ Websites/MainWebsite/SWM/src/SWM/Models/ApiModels/LocationInfoModel.cs
@@ -10,6 +10,6 @@ namespace SWM.Models.ApiModels
     {
         public string Name { get; set; }
         public string Address { get; set; }
-        public int TotalWeight { get; set; }
+        public long TotalWeight { get; set; }
     }
 }
