+++ Websites/MainWebsite/SWM/src/SWM/Models/ApiModels/ProductInfoModel.cs
@@ -9,6 +9,6 @@ namespace SWM.Models.ApiModels
     public class ProductInfoModel
     {
         public string ProductName { get; set; }
-        public int TotalWeight { get; set; }
+        public long TotalWeight { get; set; }
     }
 }
